+++ Evergreen/Public/Export-EvergreenManifest.ps1
@@ -16,9 +16,9 @@ Function Export-EvergreenManifest {
     catch {
         Write-Warning -Message "Please list valid application names with Find-EvergreenApp."
         Write-Warning -Message "Documentation on how to contribute a new application to the Evergreen project can be found at: $($script:resourceStrings.Uri.Docs)."
-        Throw "Failed to retrieve manifest for application: $Name."
+        throw "Failed to retrieve manifest for application: $Name."
     }
-    If ($Output) {
+    if ($Output) {
         Write-Output -InputObject $Output
     }
 }