+++ Evergreen/Public/Get-EvergreenApp.ps1
@@ -29,29 +29,29 @@ Function Get-EvergreenApp {
             $Function = [System.IO.Path]::Combine($MyInvocation.MyCommand.Module.ModuleBase, "Apps", "Get-$Name.ps1")
         }
         catch {
-            throw "$($MyInvocation.MyCommand): Failed to combine: $($MyInvocation.MyCommand.Module.ModuleBase), Apps, Get-$Name.ps1"
+            throw "Failed to combine: $($MyInvocation.MyCommand.Module.ModuleBase), Apps, Get-$Name.ps1"
         }
 
         #region Test that the function exists and run it to return output
         if (Test-Path -Path $Function -PathType "Leaf" -ErrorAction "SilentlyContinue") {
-            Write-Verbose -Message "$($MyInvocation.MyCommand): Function exists: $Function."
+            Write-Verbose -Message "Function exists: $Function."
 
             try {
                 # Dot source the function so that we can use it
                 # Import function here rather than at module import to reduce IO and memory footprint as the module grows
                 # This also allows us to add an application manifest and function without having to re-load the module
-                Write-Verbose -Message "$($MyInvocation.MyCommand): Dot sourcing: $Function."
+                Write-Verbose -Message "Dot sourcing: $Function."
                 . $Function
             }
             catch {
-                throw "$($MyInvocation.MyCommand): Failed to load function: $Function."
+                throw "Failed to load function: $Function."
             }
 
             try {
                 # Run the function to grab the application details; pass the per-app manifest to the app function
                 # Application manifests are located under Evergreen/Manifests
                 if ($PSBoundParameters.ContainsKey("AppParams")) {
-                    Write-Verbose -Message "$($MyInvocation.MyCommand): Adding AppParams."
+                    Write-Verbose -Message "Adding AppParams."
                     $params = @{
                         res = (Get-FunctionResource -AppName $Name)
                     }
@@ -62,27 +62,27 @@ Function Get-EvergreenApp {
                         res = (Get-FunctionResource -AppName $Name)
                     }
                 }
-                Write-Verbose -Message "$($MyInvocation.MyCommand): Calling: Get-$Name."
+                Write-Verbose -Message "Calling: Get-$Name."
                 $Output = & Get-$Name @params
             }
             catch {
-                Write-Error -Message "$($MyInvocation.MyCommand): Internal application function: $Function, failed with: $($_.Exception.Message)"
+                Write-Error -Message "Internal application function: $Function, failed with: $($_.Exception.Message)"
             }
 
             # if we get an object, return it to the pipeline
             # Sort object on the Version property
             if ($Output) {
-                Write-Verbose -Message "$($MyInvocation.MyCommand): Output result from: $Function."
+                Write-Verbose -Message "Output result from: $Function."
                 Write-Output -InputObject ($Output | Sort-Object -Property @{ Expression = { [System.Version]$_.Version }; Descending = $true } -ErrorAction "SilentlyContinue")
             }
             else {
-                throw "$($MyInvocation.MyCommand): Failed to capture output from: Get-$Name."
+                throw "Failed to capture output from: Get-$Name."
             }
         }
         else {
             Write-Warning -Message "Please list valid application names with Find-EvergreenApp."
             Write-Warning -Message "Documentation on how to contribute a new application to the Evergreen project can be found at: $($script:resourceStrings.Uri.Docs)."
-            throw "$($MyInvocation.MyCommand): Cannot find application script at: $Function."
+            throw "Cannot find application script at: $Function."
         }
         #endregion
     }